+++ packages/adminui-vue/components/UTableEntity/filter-templates/many.vue
@@ -42,7 +42,7 @@
 </template>
 
 <script>
-const FilterSubmitButton = require('../components/filter-submit-button.vue').default
+const FilterSubmitButton = require('../components/FilterSubmitButton.vue').default
 const { mapGetters, mapActions } = require('vuex')
 
 export default {