+++ packages/adminui-vue/components/UTableEntity/filter-templates/number.vue
@@ -70,7 +70,7 @@
 </template>
 
 <script>
-const FilterSubmitButton = require('../components/filter-submit-button.vue').default
+const FilterSubmitButton = require('../components/FilterSubmitButton.vue').default
 const { mapActions } = require('vuex')
 
 export default {